backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
sqlalchemy==2.0.36
asyncpg==0.30.0
aiosqlite==0.20.0
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
email-validator==2.2.0
httpx==0.28.1
greenlet==3.0.3
